Draw line from vehicle to next position setpoint
QGC already marks "Goto" locations and active waypoints, but in some
cases the vehicle will split a movement into several sequential position
waypoints. One example is a QuadChute in PX4, where the vehicle will
first backtrack to the location at which the QuadChute was triggered,
then it will increase its altitude to the configured level, and only
_then_ will it start returning to the nearest safe point. This behaviour
can be confusing to the operator.

This commit solves this issue by using the Mavlink message
UTM_GLOBAL_POSITION which has the values next_lat, next_lon and
next_alt. In PX4, this corresponds to the lat/lon/alt of
"position_setpoint_triplet.current", which is what the position
controller actually uses.

The UI element can be hidden/shown from the FlyView settings, and is
disabled by default.