geofence: disable maxAlitude for exclusion fence
PX4 will not use maxAltitude for exclusion fence, so the field is
disabled to make the UI more intuitive. Also, maxAltitude is not saved
or sent for exclusive fences. It will, however, be retained in QGC when
switching between inclusion/exclusion to avoid unintentionally setting
maxAltitude to 0 (disabled) if exclusion is selected/deselected by
mistake.