geofence: clarify altitude limits for polygons/circles
Specify AMSL for max altitude. Since "0" means disabled, this will be
rendered as "Disabled". The word "Disabled" (case insensitive) can be
entered instead of 0. Also fix validation a bit, prior a too high number
(e.g. 999999) that would be entered with intent to disable the altitude
would not be accepted, but this would be invisible, causing the last
valid number to be used. Range valid range to 0-99999 (and "Disabled"),
and show red text if the field is invalid.