Mark FW rally point as "important"
Increase the size and add a black border to it. Also mark rally points
with R, M or F according to if they are always active, MR only or FW only.

The used integer values for "Always", "MR only" and "FW only" is hard
codeed in the qml where needed. This is not ideal or pretty, but it is a
fast way of getting this feature.