Show new altitude ralatve to current altitude in slider
This is in addition to the existing new altitude in AMSL or relative to
home. Also do some minor cleanup and specify that the previous
"relative" new altitude is relative to home.

The label will not also turn red if the new altitude is more than 120 m
AGL (according to distance sensor, if available). The limit is hard-coded
for now, andshould be added to configuration later.