fix: fix segfault error when pressing stale QGCMenu
When visibleADSBVehicles signal is emitted, all VehicleMapItems are destroyed and then the updated are created. If the QGCMenu was open when the VehicleMapItem was destroyed, you would get a segfault on next click, as the refrence to the menu is destroyed. I did not manage to close the QGCMenu pragmatically on component destroy, so I switched to QGCPopupDialog